refactor mesh factory
