+++ TGC.Core/SceneLoader/TgcSceneLoader.cs
@@ -6,6 +6,7 @@ using System.IO.Compression;
 using TGC.Core.BoundingVolumes;
 using TGC.Core.Direct3D;
 using TGC.Core.Mathematica;
+using TGC.Core.MeshFactory;
 using TGC.Core.PortalRendering;
 using TGC.Core.Textures;
 
@@ -735,54 +736,5 @@ namespace TGC.Core.SceneLoader
 
         #endregion Mesh FVF
 
-        #region MeshFactory
-
-        /// <summary>
-        ///     Factory para permitir crear una instancia especifica de la clase TgcMesh
-        /// </summary>
-        public interface IMeshFactory
-        {
-            /// <summary>
-            ///     Crear una nueva instancia de la clase TgcMesh o derivados
-            /// </summary>
-            /// <param name="d3dMesh">Mesh de Direct3D</param>
-            /// <param name="meshName">Nombre de la malla</param>
-            /// <param name="renderType">Tipo de renderizado de la malla</param>
-            /// <param name="meshData">Datos de la malla</param>
-            /// <returns>Instancia de TgcMesh creada</returns>
-            TgcMesh createNewMesh(Mesh d3dMesh, string meshName, TgcMesh.MeshRenderType renderType);
-
-            /// <summary>
-            ///     Crear una nueva malla que es una instancia de otra malla original.
-            ///     Crear una instancia de la clase TgcMesh o derivados
-            /// </summary>
-            /// <param name="name">Nombre de la malla</param>
-            /// <param name="parentInstance">Malla original desde la cual basarse</param>
-            /// <param name="translation">Traslación respecto de la malla original</param>
-            /// <param name="rotation">Rotación respecto de la malla original</param>
-            /// <param name="scale">Escala respecto de la malla original</param>
-            /// <returns>Instancia de TgcMesh creada</returns>
-            TgcMesh createNewMeshInstance(string meshName, TgcMesh originalMesh, TGCVector3 translation, TGCVector3 rotation,
-                TGCVector3 scale);
-        }
-
-        /// <summary>
-        ///     Factory default que crea una instancia de la clase TgcMesh
-        /// </summary>
-        public class DefaultMeshFactory : IMeshFactory
-        {
-            public TgcMesh createNewMesh(Mesh d3dMesh, string meshName, TgcMesh.MeshRenderType renderType)
-            {
-                return new TgcMesh(d3dMesh, meshName, renderType);
-            }
-
-            public TgcMesh createNewMeshInstance(string meshName, TgcMesh originalMesh, TGCVector3 translation,
-                TGCVector3 rotation, TGCVector3 scale)
-            {
-                return new TgcMesh(meshName, originalMesh, translation, rotation, scale);
-            }
-        }
-
-        #endregion MeshFactory
     }
 }
